Convert API response JSON to Python dict
Some methods in this library are returning the API response JSON
string without converting it to a Python dict.
This is forcing the caller to selectively call json.loads
depending on the retrieved resource (e.g. `users` resource requires
json.loads, but `items` doesn't).

This patch makes all the API responses to be converted to a Python
dict, so the caller can consistently use the return value as a dict
regardless the requested resource.