Make GraphQL client return JSON response
Currently, the library converts the response.text returned by the
GraphQL client to a Python dict using json.loads.

But the requests library already has a builtin JSON decoder that
converts a JSON response to a Python dict.

This patch makes the GraphQL return the response.json() result.
The callers no longer need to convert the return value to a dict.